requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
convex==0.6.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
duffel-api==0.6.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
twilio==9.1.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
